Recyclerview partial update - 6

+++ app/src/main/java/com/hwl/beta/net/near/NearCircleService.java
@@ -144,7 +144,7 @@ public class NearCircleService {
                 .subscribeOn(Schedulers.io());
     }
 
-    public static Observable<DeleteNearCommentResponse> deleteComment(int commentId, String lastUpdateTime) {
+    public static Observable<DeleteNearCommentResponse> deleteComment(long commentId, String lastUpdateTime) {
         DeleteNearCommentRequest requestBody = new DeleteNearCommentRequest();
         requestBody.setUserId(UserSP.getUserId());
         requestBody.setCommentId(commentId);
